Rename API modules
