+++ src/adapters/fiat-exchange-rates-api.ts
